chore(core-manager): move server logic into server folder
