+++ packages/core-manager/src/server/plugins/index.ts
